Move upload annotations CSV code from the upload app to the annotations app.
