+++ project/annotations/static/js/UploadAnnotationsHelper.js
