+++ project/annotations/templates/annotations/help_upload.html
