+++ project/lib/templates/lib/nav_test.html
@@ -197,7 +197,7 @@
           <span>Metadata</span>
         </a></li>
 
-        <li><a href="{% url 'upload_annotations_csv' source.pk %}">
+        <li><a href="{% url 'annotations_upload_page' source.pk %}">
           <span>Annotations (CSV)</span>
         </a></li>
